Refactor loader, resolver, release resolver out of the model package
* `LoadManifest` is in separate `loader` package to reduce size when vendoring fissile
* downloading and opening releases is now in a separate
  `releaseresolver` package, which can be replaced when vendoring
  fissile
* extract manifest resolving/calculating into `resolver` package
* `resolver` package does most of the manifest validations now
  packages
* methods kept on structs, i.e. instance group, if possible
* kept load options in model package for simplicity
* added New*() functions
* kept tests intact
* totally avoids import cycles, had to move some tests into `_test`
  packages
* several structs and methods are public now, only because this change
  avoids to modify existing tests

Renamed and commented load release functions
* LoadReleases -> LoadReleasesFromDisk
* loadReleaseReferences -> downloadReleaseReferences

[#161744671](https://www.pivotaltracker.com/story/show/161744671)
