+++ model/instance_groups_test.go
@@ -8,7 +8,6 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 	yaml "gopkg.in/yaml.v2"
 )
 
@@ -30,28 +29,6 @@ func TestRolesSort(t *testing.T) {
 	sort.Sort(instanceGroups)
 	assert.Equal(instanceGroups[0].Name, "ccc")
 	assert.Equal(instanceGroups[1].Name, "ddd")
-}
-
-func TestGetScriptPaths(t *testing.T) {
-	workDir, err := os.Getwd()
-	assert.NoError(t, err)
-
-	torReleasePath := filepath.Join(workDir, "../test-assets/tor-boshrelease")
-	roleManifestPath := filepath.Join(workDir, "../test-assets/role-manifests/model/tor-good.yml")
-	roleManifest, err := LoadRoleManifest(roleManifestPath, LoadRoleManifestOptions{
-		ReleasePaths: []string{torReleasePath},
-		BOSHCacheDir: filepath.Join(workDir, "../test-assets/bosh-cache"),
-		ValidationOptions: RoleManifestValidationOptions{
-			AllowMissingScripts: true,
-		}})
-	assert.NoError(t, err)
-	require.NotNil(t, roleManifest)
-
-	fullScripts := roleManifest.InstanceGroups[0].GetScriptPaths()
-	assert.Len(t, fullScripts, 3)
-	for _, leafName := range []string{"environ.sh", "myrole.sh", "post_config_script.sh"} {
-		assert.Equal(t, filepath.Join(workDir, "../test-assets/role-manifests/model/scripts", leafName), fullScripts["scripts/"+leafName])
-	}
 }
 
 func TestGetScriptSignatures(t *testing.T) {
@@ -97,7 +74,7 @@ func TestGetScriptSignatures(t *testing.T) {
 		JobReferences: JobReferences{refRole.JobReferences[0], refRole.JobReferences[1]},
 		Scripts:       []string{scriptName},
 		roleManifest: &RoleManifest{
-			manifestFilePath: releasePath,
+			ManifestFilePath: releasePath,
 		},
 	}
